[ThemeBundle] Added translation integration tests
