+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/Bundle/TestBundle/Resources/translations/messages.en.yml
@@ -1,0 +1,6 @@
+test:
+  bundle: 'BUNDLE/Resources/translations'
+  app_bundle: 'BUNDLE/Resources/translations'
+  app: 'BUNDLE/Resources/translations'
+  theme_bundle: 'BUNDLE/Resources/translations'
+  theme: 'BUNDLE/Resources/translations'
