+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/TemplatingTest.php
@@ -2,12 +2,10 @@
 
 namespace Sylius\Bundle\ThemeBundle\Tests\Functional;
 
-use Symfony\Bundle\FrameworkBundle\Client;
-
 /**
  * @author Kamil Kokot <kamil.kokot@lakion.com>
  */
-class TemplatingTest extends WebTestCase
+class TemplatingTest extends ThemeBundleTestCase
 {
     /**
      * @dataProvider getBundleTemplates
@@ -61,34 +59,5 @@ class TemplatingTest extends WebTestCase
             [':Templating:bothThemesTemplate.txt.twig', ':Templating:bothThemesTemplate.txt.twig|sylius/first-test-theme'],
             [':Templating:lastThemeTemplate.txt.twig', ':Templating:lastThemeTemplate.txt.twig|sylius/second-test-theme'],
         ];
-    }
-
-    protected function setUp()
-    {
-        parent::setUp();
-
-        $this->deleteTmpDir('ThemeBundleTest');
-    }
-
-    protected function tearDown()
-    {
-        parent::tearDown();
-
-        $this->deleteTmpDir('ThemeBundleTest');
-    }
-
-    /**
-     * @return Client
-     */
-    private function getClient()
-    {
-        $client = $this->createClient(array('test_case' => 'DefaultTestCase', 'root_config' => 'config.yml'));
-        try {
-            $client->insulate();
-        } catch (\RuntimeException $e) {
-            // Don't insulate requests if not possible to do so.
-        }
-
-        return $client;
     }
 }
