+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/DefaultTestCase/themes/FirstTestTheme/TestBundle/translations/messages.en.yml
@@ -1,0 +1,3 @@
+test:
+  theme_bundle: 'THEME/BUNDLE_NAME/translations'
+  theme: 'THEME/BUNDLE_NAME/translations'
