+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/DefaultTestCase/themes/FirstTestTheme/translations/messages.en.yml
@@ -1,0 +1,2 @@
+test:
+  theme: 'THEME/translations'
