+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/DefaultTestCase/themes/SecondTestTheme/TestBundle/translations/messages.en.yml
@@ -1,0 +1,3 @@
+test:
+  theme_bundle: 'That should be overrided by first test theme.'
+  theme: 'That should be overrided by first test theme.'
