+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/DefaultTestCase/themes/SecondTestTheme/translations/messages.en.yml
@@ -1,0 +1,2 @@
+test:
+  theme: 'That should be overrided by first test theme.'
