+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/Resources/TestBundle/translations/messages.en.yml
@@ -1,0 +1,5 @@
+test:
+  app_bundle: 'app/Resources/BUNDLE_NAME/translations'
+  app: 'app/Resources/BUNDLE_NAME/translations'
+  theme_bundle: 'app/Resources/BUNDLE_NAME/translations'
+  theme: 'app/Resources/BUNDLE_NAME/translations'
