+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/Resources/translations/messages.en.yml
@@ -1,0 +1,4 @@
+test:
+  app: 'app/Resources/translations'
+  theme_bundle: 'app/Resources/translations'
+  theme: 'app/Resources/translations'
