+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/Resources/views/Translation/translationsTest.txt.twig
@@ -1,0 +1,5 @@
+BUNDLE/Resources/translations: {{ 'test.bundle'|trans }}
+app/Resources/BUNDLE_NAME/translations: {{ 'test.app_bundle'|trans }}
+app/Resources/translations: {{ 'test.app'|trans }}
+THEME/BUNDLE_NAME/translations: {{ 'test.theme_bundle'|trans }}
+THEME/translations: {{ 'test.theme'|trans }}
