+++ src/Sylius/Bundle/ThemeBundle/Tests/Functional/app/config/framework.yml
@@ -15,6 +15,7 @@ framework:
     fragments:       ~
     http_method_override: true
     test: ~
+    translator: { fallback: [en] }
 
 twig:
     debug:            false
